requirements.txt
unittest2==0.5.1
nose>=1.1.2
tox>=1.3
Sphinx>=1.2.2
coverage>=3.7.1
ordereddict>=1.1
six>=1.6.1
python-dateutil==2.2
sphinx_rtd_theme>=0.1.6
wheel>=0.24.0
pytimeparse>=1.1.5
